+++ R/mnis_general_election_results.R
@@ -1,6 +1,5 @@
-#' mnis_general_election_results
-#'
-#' Returns a list with details of the search parameter and a tibble with election results. Accepts queries on location type and name, and the start and end date to return general elections between. The API does not contain data for Norther Ireland.
+
+#' Returns an object containing list with details of the search parameter and a tibble with election results. Accepts queries on location type and name, and the start and end date to return general elections between. The API does not contain data for Norther Ireland.
 #' @param location_type The type of area to return information for. Accepts 'Country', 'Region', 'County', and 'Constituency'. Defaults to 'Country'.
 #' @param location_name The location to return data for. It can be the name of any Country, Region, County or Constituency. Defaults to 'Great Britain'.
 #' @param start_date Start date of search. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else than can be coerced to a date with \code{as.Date()}. Defaults to '1900-01-01' if no date is selected.
@@ -10,6 +9,7 @@
 #' @return Returns a list with details of the search parameter and a tibble with election results.
 #' @keywords mnis
 #' @export
+#' @seealso \code{\link{mnis_reference}}
 #' @examples \dontrun{
 #' x <- mnis_general_election_results(location_type = 'Country', location_name = 'England',
 #'                                  start_date = '2010-01-01', end_date = '2016-01-01')
@@ -40,11 +40,19 @@ mnis_general_election_results <- function(location_type = "Country", location_na
 
     got <- jsonlite::fromJSON(got, flatten = TRUE)
 
-    x <- tibble::as_tibble(got$ElectionResults)
+    x <- got$ElectionResults
+
+    x$ElectionResult <- as.tibble(x$ElectionResult)
 
     if (tidy == TRUE) {
 
-        x$ElectionResult <- mnis::mnis_tidy(x$ElectionResult, tidy_style)
+        names(x)[names(x)=="LocationInfo"] <- "location_info"
+
+        names(x)[names(x)=="ElectionResult"] <- "election_result"
+
+        x$election_result <- mnis::mnis_tidy(x$election_result, tidy_style)
+
+        x
 
     } else {