+++ R/mnis_mps_on_date.R
@@ -1,6 +1,5 @@
 
-#' mnis_mps_on_date
-#'
+#' A tibble with information on all MPs who were members of the House of Commons on the date specificed (if only date1 is included as a parameter), or on or between the two dates if both date1 and date2 are specified.
 #' @param date1 The date to return the list of mps from. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else than can be coerced to a date with \code{as.Date()}. Defaults to current system date.
 #' @param date2 An optional query parameter. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else than can be coerced to a date with \code{as.Date()}. If not NULL, the function returns a list of all MPs who were members between date2 and date1. Defaults to NULL.
 #' @param tidy Fix the variable names in the tibble to remove extra characters, superfluous text and convert variable names to a consistent style. Defaults to TRUE.
@@ -49,6 +48,14 @@ mnis_mps_on_date <- function(date1 = Sys.Date(), date2=NULL, tidy = TRUE, tidy_s
   mps <- got$Members$Member
 
   mps <- tibble::as_tibble(mps)
+
+  if(.Platform$OS.type=="windows"){
+
+    mps$MemberFrom <- stringi::stri_trans_general(mps$MemberFrom, "latin-ascii")
+
+    mps$MemberFrom <- gsub("Ynys MA\U00B4n", "Ynys M\U00F4n", mps$MemberFrom)
+
+  }
 
   if (tidy == TRUE) {