+++ NEWS.md
@@ -1,3 +1,14 @@
+# mnis 0.2.7
+
+## Bug Fixes
+
+Fixed bug in `mnis_tidy()` that didn't correct data schema names.
+
+Fixed test errors for `mnis_joined_between()`.
+
+Removed `basic_details` parameter from `mnis_extra()`, as it didn't provide any information that wasn't included by other parameters.
+
+Changed `mnis_extra()` code to eliminate duplicate columns.
 
 # mnis 0.2.6